Rearrange code and add comments
