+++ crates/shared/src/contracts.rs
@@ -1,7 +1,14 @@
 use anyhow::{anyhow, bail, Result};
+use contracts::GPv2Settlement;
 use ethcontract::{
     common::{contract::Network, DeploymentInformation},
     Contract,
+};
+use web3::types::U64;
+
+use crate::{
+    current_block::{block_number_to_block_number_hash, BlockNumberHash},
+    ethrpc::Web3,
 };
 
 pub fn deployment(contract: &Contract, chain_id: u64) -> Result<&Network> {
@@ -25,4 +32,14 @@ pub async fn deployment_block(contract: &Contract, chain_id: u64) -> Result<u64>
             bail!("missing deployment block number for {}", tx)
         }
     }
+}
+
+pub async fn settlement_deployment_block_number_hash(
+    web3: &Web3,
+    chain_id: u64,
+) -> Result<BlockNumberHash> {
+    let block_number = deployment_block(GPv2Settlement::raw_contract(), chain_id).await?;
+    block_number_to_block_number_hash(web3, U64::from(block_number).into())
+        .await
+        .ok_or_else(|| anyhow!("Deployment block not found"))
 }
